tgp-2890: move categorisation controller spec and views into folders
